LayoutTests/platform/ios/imported/w3c/web-platform-tests/selection/move-by-word-with-symbol-expected.txt

~>>>>>p+++


PASS Symbols should be included while moving backward by word
PASS Symbols should be included while moving forward by word
FAIL Symbols should be included while moving backward by word in textarea element assert_equals: Caret should move to the beginning expected 0 but got 4
FAIL Symbols should be included while moving forward by word in textarea element assert_equals: Caret should move to the end expected 8 but got 5
FAIL Symbols should be included while moving backward by word in input element assert_equals: Caret should move to the beginning expected 0 but got 4
FAIL Symbols should be included while moving forward by word in input element assert_equals: Caret should move to the end expected 10 but got 5
